check table existing before remove column on product_detail_histories table

+++ db/migrate/20160106115521_remove_cost_from_product_detail_histories.rb
@@ -1,5 +1,5 @@
 class RemoveCostFromProductDetailHistories < ActiveRecord::Migration
   def change
-    remove_column :product_detail_histories, :cost, :decimal
+    remove_column :product_detail_histories, :cost, :decimal if ActiveRecord::Base.connection.table_exists? 'product_detail_histories'
   end
 end
